+++ src/main/java/org/foo/modules/jahia/strava/utils/RequestUtils.java
@@ -11,10 +11,10 @@ public final class RequestUtils {
 
     public static final String SESSION_REQUEST_URI = "my.request_uri";
     public static final String TOKEN_EXPIRES_AT = "expiresAt";
-    public static final long TOKEN_MIN_LIFE_IN_SECONDS = 60L;
+    public static final int TOKEN_MIN_LIFE_IN_SECONDS = 60;
 
     public static Map<String, Object> addExpireAtInTokenData(Map<String, Object> tokenData) {
-        tokenData.put(TOKEN_EXPIRES_AT, (System.currentTimeMillis() / 1000L) + (Integer) tokenData.get(JahiaOAuthConstants.TOKEN_EXPIRES_IN));
+        tokenData.put(TOKEN_EXPIRES_AT, (System.currentTimeMillis() / 1000) + (Integer) tokenData.get(JahiaOAuthConstants.TOKEN_EXPIRES_IN));
         return tokenData;
     }
 }
